requirements.txt
asgiref==3.7.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
distlib==0.3.8
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.6
msgpack==1.0.8
oauth2client==4.1.3
pillow==10.3.0
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-jwt==4.1.0
pytz==2024.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.25.0
wheel==0.42.0
